requirements.txt
pandas>=1.3.0
numpy>=1.21.0
python-dateutil>=2.8.0
scikit-learn>=0.24.0
chardet>=4.0.0
scipy>=1.7.0
drain3>=0.9.0
psycopg2-binary